Adding Python code to launch BlackBerry browser automatically
The BlackBerry10 browser doesn't launch automatically with WebDriver, which
makes automated testing problematic. This patch enables the automatic
launching of the BlackBerry10 browser, if the user has the proper env.
setup.

This patch assumes the IP/port of the device. The user must provide location
of required tools and device password.

Code review updates:
* making BB tools easier
* Fixed up python styling issues (pep8). Added a better description of the args.
* Realized that blackberry tools when installed with the WebWorks SDK are in the path
so changed it so those are handled by default. Also made the 'typical' ip address
used to be the default, now the user only needs to supply a device password if
everything is setup in the standard way.
* Added some directions/hints on getting the webworks SDK, location should never
change.
* Removing Wait(3) + improving errors
* I found a better way to wait for the BB10 browser to launch, so
it now polls until the browser launches and is ready.
* Improved error messaging to give more details for end user.
* Switching to WebDriverWait
last patch had the potential to get stuck in an infinite loop. Updated
code to use WebDriverWait to watch for the browser to load. Also lead to
much cleaner code.
* Using LOAD_TIMEOUT for easier customization
* Had a CONTS created but never used it. Using it now, reduced to 5s
as that still consistently works.
* Fixing Stackswallowing + nits
* Stack is now properly passed along with an exception. Added message when loading
browser

Signed-off-by: Luke Inman-Semerau <luke.semerau@gmail.com>