Support GeckoDriver addon install/uninstall commands
Adds support for the following commands introduced in GeckoDriver 0.17:

  * POST /session/{session id}/moz/addon/install
  * POST /session/{session id}/moz/addon/uninstall

For Ruby, implemented as a role-based interface (though I'm not sure
it was really needed in this case since it's unlikely that any other
driver will ever support these commands).

For Python, implemented directly in Firefox webdriver.

Tests are missing because all the extension
files available for Selenium build (legacy Firefox driver extension,
Firebug in third_party) fail to install on latest versions of Firefox.

Related to #4215.